lil-notification/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aioredis==1.0.0           # via channels-redis
asgiref==2.3.0            # via channels, channels-redis
asn1crypto==0.24.0        # via cryptography
async-timeout==2.0.1      # via aioredis, asgiref
attrs==17.4.0             # via automat, pytest
autobahn==18.4.1          # via daphne
automat==0.6.0            # via twisted
bcrypt==3.1.4             # via paramiko
cffi==1.11.5              # via bcrypt, cryptography, pynacl
channels-redis==2.1.1
channels==2.1.0
click==6.7                # via pip-tools
constantly==15.1.0        # via twisted
cryptography==2.2.2       # via paramiko
daphne==2.1.0             # via channels
django-filter==1.1.0
django-simple-history==2.0
django==2.0.4
djangorestframework==3.8.2
fabric3==1.14.post1
first==2.0.1              # via pip-tools
flake8==3.5.0
hiredis==0.2.0            # via aioredis
hyperlink==18.0.0         # via twisted
idna==2.6                 # via cryptography, hyperlink
incremental==17.5.0       # via twisted
mccabe==0.6.1             # via flake8
more-itertools==4.1.0     # via pytest
msgpack==0.5.6            # via channels-redis
paramiko==2.4.1           # via fabric3
pip-tools==2.0.1
pluggy==0.6.0             # via pytest
psycopg2-binary==2.7.4
py==1.5.3                 # via pytest
pyasn1==0.4.2             # via paramiko
pycodestyle==2.3.1        # via flake8
pycparser==2.18           # via cffi
pyflakes==1.6.0           # via flake8
pynacl==1.2.1             # via paramiko
pytest-django==3.2.1
pytest==3.5.0             # via pytest-django
pytz==2018.4              # via django
six==1.11.0               # via autobahn, automat, bcrypt, cryptography, fabric3, more-itertools, pip-tools, pynacl, pytest, txaio
twisted==17.9.0           # via daphne
txaio==2.10.0             # via autobahn
whitenoise==3.3.1
zope.interface==4.4.3     # via twisted
